Modify Delete : confirm before delete & Modify ArticleController.doDelete

+++ src/webapp/jsp/usr/article/list.jsp
@@ -9,7 +9,7 @@
 <script src="https://cdn.tailwindcss.com"></script>
 
 <section>
-    <div class = "container mx-auto">
+    <div class = "container px-3 mx-auto">
         <h1 class = "font-bold text-lg">게시물 리스트</h1>
 
         <ul class = "mt-5">
@@ -17,7 +17,7 @@
             <li class = "flex">
                 <a class="w-[50] hover:underline hover:text-[blue]" href="/usr/article/detail/free/<%=articleDto.getId()%>"><%=articleDto.getId()%></a>
                 <a class="flex-grow hover:underline hover:text-[blue]" href="/usr/article/detail/free/<%=articleDto.getId()%>"><%=articleDto.getTitle()%></a>
-                <a class="w-[100] hover:underline hover:text-[red]" href="/usr/article/delete/free/<%=articleDto.getId()%>">삭제</a>
+                <a onclick="if(!confirm('정말로 삭제하시겠습니까?')) return false;" class="w-[100] hover:underline hover:text-[red]" href="/usr/article/delete/free/<%=articleDto.getId()%>">삭제</a>
             </li>
             <% } %>
         </ul>
